feat: separate bar component
